fix(abi): batch 1 — syscall dispatch, signal frame, dev_t, mmap compat
ABI compatibility fixes from code review (F2/F7/F11):

- F11-01: Syscall 121/122/123 mapping corrected (getparam/setaffinity/
  getaffinity were rotated)
- F11-02: sys_close errno sign-extension fixed (u32→u64 zero-extended
  instead of sign-extending negative error codes)
- F11-03: Removed MAP_ANONYMOUS forced PROT_READ|PROT_WRITE override
  (POSIX requires PROT_NONE anonymous mappings to be valid)
- F2-01: UContext field order fixed to match Linux UAPI: uc_flags,
  uc_link, uc_stack, uc_sigmask, padding, uc_mcontext
- F2-02: SigContext restructured to sc_regs[32] matching Linux
  user_regs_struct: [0]=pc, [1]=ra, ..., [31]=t6
- F2-03: SS_ONSTACK=1, SS_DISABLE=2 swapped to match Linux values
- F7-01: DevNo encoding changed from (major<<32)|minor to
  (major<<20)|minor matching Linux kernel internal dev_t format

F11-04/05/06 marked as false positives — mmap_error constants are
already negative i64.

+++ kernel/src/syscall/file.rs
@@ -142,7 +142,7 @@ pub fn sys_close(args: SyscallArgs) -> u64 {
     unsafe {
         match close_file_fd(fd) {
             Ok(()) => 0,
-            Err(e) => e as u32 as u64,
+            Err(e) => (e as i64) as u64,
         }
     }
 }
